+++ geeadd/ee_del_meta.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 
 __copyright__ = """
 
-    Copyright 2019 Samapriya Roy
+    Copyright 2024 Samapriya Roy
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -18,8 +18,9 @@ __copyright__ = """
 
 """
 __license__ = "Apache 2.0"
-import ee
 import os
+
+import ee
 
 
 def delprop(collection_path, property):